+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### 12.56.0
+
+* `[fmtc]` Added named colors support
+* `[system]` Fixed fuzz tests
+
 ### 12.55.2
 
 * `[fmtc]` Reverted changes made in 12.55.1